# app/static/swagger.yaml
openapi: 3.0.0
info:
  title: Authentication, Company, Job, Statistics, and User API
  description: API for managing user authentication, company records, job postings, statistical data, and user interactions.
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: User Logout
      description: Logs out a user by invalidating the provided refresh token.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: The refresh token to invalidate.
      responses:
        200:
          description: User logged out successfully.
        400:
          description: Refresh token is missing or invalid.
        500:
          description: Internal server error.
  /api/auth/signup:
    post:
      tags:
        - Auth
      summary: User Signup
      description: Creates a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: ["admin", "employer", "applicant"]
                company_name:
                  type: string
                  description: Required if the role is employer.
                company_link:
                  type: string
                  description: Required if the role is employer.
      responses:
        201:
          description: User created successfully.
        400:
          description: Validation error.
        500:
          description: Internal server error.
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: User Login
      description: Logs in a user and provides access and refresh tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        200:
          description: Login successful.
        401:
          description: Invalid credentials.
        500:
          description: Internal server error.
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Token Refresh
      description: Refreshes the access token using a valid refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  description: The refresh token to use for generating a new access token.
      responses:
        200:
          description: New access token generated.
        403:
          description: Token is blacklisted.
        401:
          description: Invalid or expired refresh token.
        500:
          description: Internal server error.
  /api/companies/:
    get:
      tags:
        - Company
      summary: List all companies
      description: Fetch a list of all companies.
      responses:
        200:
          description: List of companies fetched successfully.
        500:
          description: Internal server error.
    post:
      tags:
        - Company
      summary: "Add a new company"
      description: "Create a new company record."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "Name of the company."
                link:
                  type: string
                  description: "Link to the company website."
      responses:
        201:
          description: "Company created successfully."
        400:
          description: "Validation error."
        500:
          description: "Internal server error."
  /api/companies/{company_id}:
    get:
      tags:
        - Company
      summary: Get company details
      description: Fetch details of a specific company by its ID.
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: integer
          description: ID of the company to fetch.
      responses:
        200:
          description: Company details fetched successfully.
        404:
          description: Company not found.
        500:
          description: Internal server error.
    put:
      tags:
        - Company
      summary: "Update company details"
      description: "Update details of an existing company."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: integer
          description: "ID of the company to update."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: "New name of the company."
                link:
                  type: string
                  description: "New link to the company website."
      responses:
        200:
          description: "Company updated successfully."
        400:
          description: "Validation error."
        500:
          description: "Internal server error."
    delete:
      tags:
        - Company
      summary: "Delete a company"
      description: "Delete a company by its ID."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: company_id
          required: true
          schema:
            type: integer
          description: "ID of the company to delete."
      responses:
        200:
          description: "Company deleted successfully."
        400:
          description: "Validation error."
        500:
          description: "Internal server error."
...